Vista comentarios

+++ backend/src/main/java/com/photoshare/repository/UsuarioRepository.java
@@ -2,10 +2,9 @@ package com.photoshare.repository;
 
 import com.photoshare.model.Usuario;
 import org.springframework.data.jpa.repository.JpaRepository;
-
 import java.util.Optional;
 
-public interface UsuarioRepository extends JpaRepository<Usuario, Long> {
+public interface UsuarioRepository extends JpaRepository<Usuario, Integer> {
     Optional<Usuario> findByNombreUsuario(String nombreUsuario);
     Optional<Usuario> findByCorreo(String correo);
 }
