+++ backend/src/main/java/com/photoshare/service/UsuarioService.java
@@ -17,27 +17,22 @@ public class UsuarioService {
 
   private final UsuarioRepository usuarioRepository;
 
-  // public List<UsuarioDTO> obtenerTodos() {
-  // return usuarioRepository.findAll().stream()
-  // .map(UsuarioDTO::fromEntity)
-  // .collect(Collectors.toList());
-  // }
-
-  public List<UsuarioDTO> obtenerTodosExcluyendo(Long idActual) {
+  // Devuelve todos los usuarios excepto el actual
+  public List<UsuarioDTO> obtenerTodosExcluyendo(Integer idActual) {
     return usuarioRepository.findAll().stream()
         .filter(usuario -> !usuario.getIdUsuario().equals(idActual))
         .map(UsuarioDTO::fromEntity)
         .collect(Collectors.toList());
   }
 
-  public void eliminarPorId(Long id) {
+  public void eliminarPorId(Integer id) {
     if (!usuarioRepository.existsById(id)) {
       throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Usuario no encontrado");
     }
     usuarioRepository.deleteById(id);
   }
 
-  public void actualizarUsuario(Long id, UsuarioDTO dto) {
+  public void actualizarUsuario(Integer id, UsuarioDTO dto) {
     Usuario usuario = usuarioRepository.findById(id)
         .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Usuario no encontrado"));
